issue55: remove DTO validators
