--- MultCo-ISD-API/MultCo-ISD-API/V1/Validators/DivisionV1DTOValidator.cs
@@ -1,16 +1,0 @@
-﻿using FluentValidation;
-using MultCo_ISD_API.V1.DTO;
-
-namespace MultCo_ISD_API.V1.Validators
-{
-    public class DivisionValidator : AbstractValidator<DivisionV1DTO>
-    {
-        public DivisionValidator()
-        {
-            RuleFor(x => x.DivisionID).NotNull();
-            RuleFor(x => x.DivisionCode).NotNull();
-            RuleFor(x => x.DivisionName).NotNull().MaximumLength(20);
-
-        }
-    }
-}
