--- MultCo-ISD-API/MultCo-ISD-API/V1/Validators/LanguageV1DTOValidator.cs
@@ -1,16 +1,0 @@
-﻿using FluentValidation;
-using MultCo_ISD_API.V1.DTO;
-
-namespace MultCo_ISD_API.V1.Validators
-{
-    public class LanguageValidator : AbstractValidator<LanguageV1DTO>
-    {
-        public LanguageValidator()
-        {
-            RuleFor(x => x.LanguageID).NotNull();
-
-            RuleFor(x => x.LanguageName).NotNull().MaximumLength(10);
-
-        }
-    }
-}
