+++ MultCo-ISD-API/MultCo-ISD-API/V1/Validators/LanguageValidator.cs
@@ -1,7 +1,7 @@
 ﻿using FluentValidation;
 using MultCo_ISD_API.Models;
 
-namespace MultCo_ISD_API.Validators
+namespace MultCo_ISD_API.V1.Validators
 {
     public class LanguageValidator : AbstractValidator<Language>
     {
