+++ MultCo-ISD-API/MultCo-ISD-API/V1/Validators/ServiceLanguageAssociationValidator.cs
@@ -3,7 +3,7 @@ using FluentValidation;
 using FluentValidation.Validators;
 using MultCo_ISD_API.Models;
 
-namespace MultCo_ISD_API.Validators
+namespace MultCo_ISD_API.V1.Validators
 {
     public class ServiceLanguageAssociationValidator : AbstractValidator<ServiceLanguageAssociation>
     {
