+++ MultCo-ISD-API/MultCo-ISD-API/V1/Validators/ServiceLocationAssociationValidator.cs
@@ -1,16 +1,16 @@
 ﻿using System.Data;
 using FluentValidation;
 using FluentValidation.Validators;
-using MultCo_ISD_API.V1.DTO;
+using MultCo_ISD_API.Models;
 
 namespace MultCo_ISD_API.V1.Validators
 {
-    public class ServiceLocationAssociationV1DTOValidator : AbstractValidator<ServiceLocationAssociationV1DTO>
+    public class ServiceLocationAssociationV1DTOValidator : AbstractValidator<ServiceLocationAssociation>
     {
         public ServiceLocationAssociationV1DTOValidator()
         {
-            RuleFor(x => x.ServiceID).NotNull();
-            RuleFor(x => x.LocationID).NotNull();
+            RuleFor(x => x.ServiceId).NotNull();
+            RuleFor(x => x.LocationId).NotNull();
         }
     }
 }
