+++ MultCo-ISD-API/MultCo-ISD-API/V1/Validators/ServiceValidator.cs
@@ -1,11 +1,11 @@
 ﻿using System.Data;
 using FluentValidation;
 using FluentValidation.Validators;
-using MultCo_ISD_API.V1.DTO;
+using MultCo_ISD_API.Models;
 
 namespace MultCo_ISD_API.V1.Validators
 {
-    public class ServiceValidator : AbstractValidator<ServiceV1DTO>
+    public class ServiceValidator : AbstractValidator<Service>
     {
         public ServiceValidator()
         {
