--- MultCo-ISD-API/MultCo-ISD-API/Validators/ServiceLocationAssociationValidator.cs
@@ -1,16 +1,0 @@
-﻿using System.Data;
-using FluentValidation;
-using FluentValidation.Validators;
-using MultCo_ISD_API.Models;
-
-namespace MultCo_ISD_API.Validators
-{
-    public class ServiceLocationAssociationV1DTOValidator : AbstractValidator<ServiceLocationAssociation>
-    {
-        public ServiceLocationAssociationV1DTOValidator()
-        {
-            RuleFor(x => x.ServiceId).NotNull();
-            RuleFor(x => x.LocationId).NotNull();
-        }
-    }
-}
